v0.18.12 General Update
- Moved the filter player logic into the banner, it looks rough, but will be fixed.

NOTE: This adds a new layer of complexity to the filtering. It is now possible to get filter in a way that return zero games. This does not break the site, but a solution should be implemented for this. Adding the same logic that years have or making the years dependent on player is also not ideal due to both then being useStates... Choices, choices, choices...